# app/cache/test/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Trailburning Journey API",
        "contact": {
            "name": "Matt Allbuery",
            "url": "http://www.trailburning.com/",
            "email": "matt@trailburning.com"
        },
        "version": "2.0"
    },
    "basePath": "/v2",
    "consumes": [
        "application/json",
        "multipart/form-data",
        "application/x-www-form-urlencoded"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/events/{id}/assets": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Find assets by event",
                "description": "Returns all assets of a event.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Asset"
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Add an asset",
                "description": "Adds an asset fo an event.",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event the asset belongs to",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "The name of the asset",
                        "type": "string"
                    },
                    {
                        "name": "about",
                        "in": "formData",
                        "description": "About the asset",
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "formData",
                        "description": "The asset category ID",
                        "type": "string"
                    },
                    {
                        "name": "position",
                        "in": "formData",
                        "description": "The sort position, is handled automatically if not specified",
                        "type": "integer"
                    },
                    {
                        "name": "credit",
                        "in": "formData",
                        "description": "The credit of the asset",
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation. The Location header contains a link to the new event.",
                        "headers": {
                            "location": {
                                "description": "Link to the new event.",
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data."
                    }
                }
            }
        },
        "/assets/{id}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Find an asset by ID",
                "description": "Returns a single asset.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the asset to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Asset"
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Assets"
                ],
                "summary": "Update an asset",
                "description": "Updates an asset.",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the asset to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "The name of the asset",
                        "type": "string"
                    },
                    {
                        "name": "about",
                        "in": "formData",
                        "description": "About the asset",
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "formData",
                        "description": "The label of the asset category",
                        "type": "string"
                    },
                    {
                        "name": "position",
                        "in": "formData",
                        "description": "The sort position, is handled automatically if not specified",
                        "type": "integer"
                    },
                    {
                        "name": "credit",
                        "in": "formData",
                        "description": "The credit of the asset",
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid data."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Delete an asset",
                "description": "Deletes the asset.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the asset",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Journey not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Asset": {
            "required": [
                "id",
                "about",
                "category"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "about": {
                    "type": "string"
                },
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Media"
                    }
                },
                "category": {
                    "$ref": "#/definitions/AssetCategory"
                },
                "credit": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Asset"
            }
        },
        "AssetCategory": {
            "required": [
                "name",
                "label"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "AssetCategory"
            }
        },
...
        "Media": {
            "required": [
                "id",
                "mimeType",
                "path"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "metadata": {}
            },
            "xml": {
                "name": "Media"
            }
        },
...
    }
}